// packages/registry/src/forme/blocks/invoice-classic/invoice-classic.types.ts
import type {
  InvoiceLineItem,
  InvoiceParty,
  InvoicePaymentTerms,
  InvoiceSummary,
} from '../shared/invoice.types.ts';

export type InvoiceClassicData = {
  invoiceNumber: string;
  invoiceDate: string;
  dueDate: string;
  companyName: string;
  subtitle: string;
  companyAddress: string;
  companyEmail: string;
  logo?: string | undefined;
  billTo: InvoiceParty;
  items: InvoiceLineItem[];
  summary: InvoiceSummary;
  paymentTerms: InvoicePaymentTerms;
  notes?: string | undefined;
};